menambahkan data sandar dan tolak pada clearance / perjalanan

+++ backend/controller/agenController.js
@@ -72,8 +72,16 @@ const deleteAgen = async (req, res) => {
     try {
         let agenData = await agen.findOne({
             where: { id_agen: req.params.id },
-            attributes: ['nama_agen']
+            attributes: ['nama_agen', 'createdAt']
         })
+
+        let agenDate = new Date(agenData.createdAt)
+        let now = new Date()
+
+        let dateDifference = Math.floor((now - agenDate) / (1000 * 60 * 60 * 24))
+
+        if (dateDifference > 10) return res.status(500).json({ msg: "data tidak bisa dihapus" })
+
         let result = await agen.destroy({ where: { id_agen: req.params.id } })
 
         if (result == 0) return res.status(500).json({ msg: "data tidak ditemukan" })
