+++ backend/controller/kecamatanController.js
@@ -76,8 +76,16 @@ const deleteKecamatan = async (req, res) => {
     try {
         let kecamatanData = await kecamatan.findOne({
             where: { id_kecamatan: req.params.id },
-            attributes: ['nama_kecamatan']
+            attributes: ['nama_kecamatan', 'createdAt']
         })
+
+        let kecamatanDate = new Date(kecamatanData.createdAt)
+        let now = new Date()
+
+        let dateDifference = Math.floor((now - kecamatanDate) / (1000 * 60 * 60 * 24))
+
+        if (dateDifference > 10) return res.status(500).json({ msg: "data tidak bisa dihapus" })
+
         let result = await kecamatan.destroy({where: {id_kecamatan: req.params.id}})
         
         if (result == 0) return res.status(500).json({msg: "data tidak ditemukan"})
